+++ src/components/TimeSelector/HourSections/SectionText/SectionText.style.ts
